Updated openstack/openstack Project: openstack/nova  9d77ab1290f19973655fdfbc35e9770f9adaffd4
Fix logic for checking if az can be updated

The existing logic was overly complicated and missed out on those hosts
whose services were disabled.

This is a complete rewrite that makes use of a single aggregate query,
thereby bypassing a lot of the extra logic needed by the old code.

Fixes an issue in objects.AggregateList.get_by_metadata_key() where
filtering by an empty list of hosts will return metadata for all hosts.

Also removes a call to db.aggregate_metadata_get_by_metadata_key() to
avoid the need for special handling due to that method returning
metadata as sets of values instead of strings and also because the same
metadata is already fetched in another method.

Change-Id: I514e63ce863f2c77dcd47af3e3674019033a77de
Closes-Bug: #1419115