Updated openstack/openstack Project: openstack/tempest  80f4e186273f677e5dd3a8dfa0ed8813cf1846df
Switch all uses of httplib to import from six

As part of enabling python3 support in tempest we need to be able to
handle httplib usage in both python2 and python3. Six provides a compat
layer for doing this, so this commit moves all uses of httplib to get
it through six.

Change-Id: Ibf1d87908fe4004b37320439386abe5b1dc4bb51