Updated openstack/openstack Project: openstack/nova  9da050abe5490898121aa155894de448abefd1d8
view hypervisor details rest api should be allowed for non-admins

The current call to show hypervisor is broken and expects the
user to have admin privileges always.It does not allow non-
admin users to view hypervisor details. This should be
controlled based on the rule defined in the policy.json rather
than have the permissions mandated programmatically. This bug
is caused due to regression and was working earlier.So the
changes proposed here is merely retaining the older behavior.
To ensure that this works as before, the context passed to the
db api layer is elevated so that it passes the requires_admin_
context check that is present in the db layer.

Change-Id: I50437788aac72608998cf9c24d4d562b1ff20301
Closes-Bug: 1447084