Updated openstack/openstack Project: openstack/nova  e4e74150b26ee890ea6379939c9256b06626e350
Don't query compute_node through service object in nova-manage

The relationship between compute_node and service tables was
removed in a recent commit and to be able to remove compat code in
the objects, calls to get them should be done directly instead of
going through one or the other.

Since we are really just interested in the compute_node object here
this in addition saves us one extra db call.

Change-Id: I5270882321f7d16f993746498f751f17b0cb52fd
Related-Bug: #1438189