Updated openstack/openstack Project: openstack/nova  326ebe1aad6aba258d028cb7b5d49cbc7f691457
Add support for forcing migrate_flavor_data

In some cases it might be helpful to force the migration of flavor
data. For example, when operating on a snapshot of a production
database there may be instances in an operating state that would
stop the migration from finishing. Blocking on this will make it
impossible to finish the migrations (once the migrations require
all the data has been migrated).

Change-Id: I4f6c0679b144897f2bfc2fa06bc7c72b25a85fd3