Updated openstack/openstack Project: openstack/heat  d644ea964d8ad6a950d074087a482e6c1520d994
Use ipv4 subnet in Heat scenario tests

Previously we used first subnet from network and it was correct, because
there was only one subnet in network by default. According to the last changes
in devstack new ipv6 subnet was added to the all default networks.
This patch adds function to the base class, which allows to choose
subnet based on ip version. All scenario tests were updated to use only ipv4
subnet.

Also was added property for FIP, which specify ip of vip.

Change-Id: I199075909b97fe98e1a3c1d175b0b4cf7100403d
Closes-Bug: #1449896