Updated openstack/openstack Project: openstack/nova  8ba3e67c567a8a5634d0b901895dcd69af328103
Add 'suspended' lifecycle event

Hyper-V can emit events signaling that an instance was suspended.
The issue is that at the moment, there is no lifecycle event defined
for this instance state in nova.virt.event.

This patch adds this lifecycle event and ensures that the manager
will properly handle it.

Partially fixes bug: #1428505

Change-Id: I0fad28025785e77ef32b88ff813f39000d9f9e35