Updated openstack/openstack Project: openstack-infra/elastic-recheck  2f93b391487ac3c089d872a0156cef80293f616d
actually actually make window location work

The last change failed to account for the fact that bug is a hash, and
hence was not building the correct anchors ever. This fixes that.

This was also actually actually tested using the new test_server.py,
which makes live testing these changes much much better.

Change-Id: Id340bdc3e2046dde7273bccc7fe97f42e1db07b9