Updated openstack/openstack Project: openstack/python-neutronclient  b5f1b4a277914df3fffded27e1d6e74fc66f8423
Fix Python client library for Neutron

To create a subnet, we used to require positional cidr value.
With subnetpool, cidr is now optional value. With optional
positional cidr, the cidr value cannot be parsed into known_args
if it is separated from network value.

To fix the problem, we need to find cidr value from value_specs,
and use it if we find it. Cidr is optional, may not exist.

Change-Id: Ic4bccf4e7384cc891cef71fb7ce3b52fbf997345
Closes-Bug:#1442771