Updated openstack/openstack Project: openstack-infra/shade  c4145aaa5bb6fd7c7cc97bd245dcd889a75406f2
Fix functional tests to run against live clouds

It's possible a developer may have a test account on a live cloud,
rather than a devstack. To support this, we should not make assumptions
about content put into a cloud by devstack, but should request things we
want - such as "the smallest flavor you have".

Keep the attempt to use the cirros image to continue to support
devstack, but fall back to the first ubuntu image found, since that's
the most likely to be cached on compute hosts on public clouds.

Change-Id: I6f0336d02cf924b3657fdc42bd151d11d5e3a6be