Updated openstack/openstack Project: openstack/trove  d3c9b30f9eb5d5cc8fc9830aa431b68957676218
Fix Mongo status check to work for Mongo 3.0

As of version 3.0, MongoDB no longer reports "connected as:..." in
the output of mongostat. This causes the guest agent status checking
function to not work. This changes it to not grep for connected but
just check the return code from the mongostat call as it will return
non-zero if mongo isn't running or reachable.

Change-Id: If1329a5338ecaffd11919183eda640ee56635794
Closes-bug: 1448185