Updated openstack/openstack Project: openstack-infra/shade  87000d9f768f5ca95fea2727ea2ca0f32865d923
Allow for int or string ID comparisons

Some IDs are clearly strings (e.g., UUIDs) and some look like integers
(e.g., flavor ID of '100'). Let's make it easy for a user by allowing
for string or int. For example:

    flavor = cloud.get_flavor(100)

or

    flavor = cloud.get_flavor('100')

The former example using the int doesn't work, but this fixes that.

Change-Id: If201352a9917d7c8fba7b386b38fced9949f00a9