Updated openstack/openstack Project: openstack/oslo.messaging  cca84f66d49114ce4ae85af4bbf03a14bda79121
rabbit: Remove unused stuffs from publisher

The publisher code is over engineered, it allows to override
everything, but this is never used.

None of the child Class have the same signature, sometimes
the constructor use the parameter name as the parent class but for
a different purpose, that make the code hard to read.

It's was never clear which options is passed to the queue and the
exchange at this end to kombu.

This changes removes all of that stuffs, and only use the kombu
terminology for publisher parameters.

Change-Id: I3cebf3ed1647a3121dcf33e2160cf315486f5204