Updated openstack/openstack Project: openstack/nova  8a688009c8acca3fbc5c7aad75f9d765bf9d8bed
Update self with db result in InstanceInfoCache.save

The InstanceInfoCache save method calls db.instance_info_cache_update
to update the database and receives an updated sqlalchemy model as a
return value. However, the result isn't saved to the self.fields of
the InstanceInfoCache object, though it is sent to update cells,
which is inconsistent. It doesn't seem critical as we expect to
receive mostly the same data back from the database as we already
have, except for the "created_at" and "updated_at" fields.

Change-Id: I64be855a194990f898ce4e572f42975a07bf1d3c