Updated openstack/openstack Project: openstack-infra/project-config  882e79a2b2407350dee8e39e099e8799e125fe03
Add tripleo-core to puppet-tripleo

Adds tripleo-core group to have rights on puppet-tripleo project. The
rights of puppet-manager-core group are kept. This project contains
Puppet code, but is solely used from the TripleO installer, so sometimes
it's not easy to find puppet-manager-core folks to do reviews for this
project to allow dependent changes in TripleO to land. We'd like to
share the permissions on this repository between the two core groups to
expand the number of interested core reviewers and allow an easier
progress on the TripleO Puppet work. The gerritbot is already set up to
notify both IRC channels.

Change-Id: I1f91f851b5800fdf69697fdd88fbd9e41349f6d1