Updated openstack/openstack Project: openstack/keystone  f029195f250cd4f6c6cd1585fc99a46b9fb82b9e
Add openstack_project_domain to assertion

Currently, a keystone IdP does not provide the domain of the project
when generating SAML assertions. Since it is possible to have two
projects with the same name but in different domains, this patch
adds an additional attribute called "openstack_project_domain"
in the assertion to identify the domain of the project.

Closes-Bug: 1442343
bp assertion-extra-attributes

Change-Id: I62ed73d87f268c73294738845421deb87088326b