Updated openstack/openstack Project: openstack/cinder  b45528ee5248954b8c259c0cecab62d7ec7e97e9
Add retry to lvm delete

Seems we have another issue related to lvm and
snapshots, but this time it's on the delete side.

We used a simple retry mechanism for snapshot create
here:
  https://review.openstack.org/#/c/149360/8

I'm hesitant to just add another retry without looking
at dm issues, but this should address the problem
temporarily and give us a chance to collect some
data on the issue.

Change-Id: I37607a09e2f9af6f2f32ddfaa8567400541e533a
Closes-Bug: #1450110