Updated openstack/openstack Project: openstack-infra/project-config  d29682ce9b61df792056dab1870c855f90eaea3b
Correct the tag merge script logic

The previous iteration did not take into account that the most
recent tag on master is likely to have been a release candidate, and
therefore have the same major+minor value as the incoming release.
Also tags pushed to non-stable branches should be skipped.

Change-Id: I339df1d88ddaedb193b1b246a2c21a433c11495c