Updated openstack/openstack Project: openstack/governance  00afb41bc2ffc121ee3162a73a1f69217deb3d63
Update diversity tag to be a fixed 6 month window

Use stackalytics start_date option along with setting release to all to
collect data for the past 6 months.

http://stackalytics.readthedocs.org/en/latest/userdoc/api_v1.0.html

Update note at the bottom of the page to point to validate_tags now that
it is in tree and used to apply the team:diverse-affiliation tag.

Change-Id: Id60feb69704e0569873108a18a39e7787c7ab4eb