Updated openstack/openstack Project: openstack/heat  5eb30dce845cb3209f21abc1a6be9a3c64f99f70
Remove the is_id boolean check for domains

The _stack_domain_is_id is a boolean that tells you if the value in
self.stack_domain is a name or id value.

This is clunky. Replace it with just saving the id or name into their
own variables as None values in auth plugins are ignored.

Change-Id: I9f6ec6b414174870317aa167e3c953c6844ccab9