Updated openstack/openstack Project: openstack/tempest  a61db2c79917a4121792fcd3c41461d62bb384bd
Create a new project for trusts tests

The trusts tests were originally being done in the project that the admin
client is scoped to. Change this to creating a new project as this will fail if
the admin client is scoped to a domain instead of a project.

Also specify domain information when creating users and projects. If not
specified the default domain for create_user and create_project is 'default',
however if not specified at authentication time the default user domain is
CONF.identity.admin_domain_name - which may not be default.

Change-Id: I4ab3383f48894d00c1b59abdb97ab483e6ca5270