Updated openstack/openstack Project: openstack/oslo.messaging  77f952a1f7d87c35fb5d7eb7e052c5b9def59b22
rabbit: Set timeout on the underlying socket

They are some case where the underlying can be stuck
until the system socket timeout is reached, but in oslo.messaging
we very often known that is not needed to wait for ever because
the upper layer (usualy the application) expect to return after
a certain period.

So this change set the timeout on the underlying socket each we can
determine that is not needed to wait more.

Closes-bug: #1436788
Change-Id: Ie71ab8147c56eaf672585da107bec8b22af9da6c