Updated openstack/openstack Project: openstack-infra/system-config  d9375caa7e0dac282c87e9e26e7ecaeec4d496bd
Add a centos-6 node label

These are the "centos6" corollary to devstack-.* node labels, but
we're not likely to run DevStack on them. Avoid confusion by going
ahead and using the new platform naming pattern settled on in the
2015-03-31 Infra meeting.

http://eavesdrop.openstack.org/meetings/infra/2015/infra.2015-03-31-19.00.html

Also update the nodepool logging config to get it back in sync with
reality.

Change-Id: Ie629b6b8e8f2401e3769a7b2c54844d950d9fd60