Updated openstack/openstack Project: openstack/requirements  1f8f4584b77a96898bd80ded085bd2698f5ec9d2
relax final pins

The following packages were pinned to a specific version, however they
are not the kind of software that should be pinned (they aren't
linters). Being non linters means we should either work with a range,
or we should remove them from being a dependency because they have
terrible compatibility stories.

Change-Id: I62f6ab9c34300482256b34456bb1b133d8fc45b3