Updated openstack/openstack Project: openstack/nova  2e1251584f8a89e5096fbdd909072a84fb22636e
Add ability to inject routes in interfaces.template

This change adds the ability to inject routes in interfaces.template
by injecting them in the template engine.

It adds support for injecting extra routes when instances are spawned
on networks that do not have DHCP or Neutron routers.

Routes are now found in each interface in the interfaces variable.

The following attributes are available for each route:
  * gateway (eg. 192.168.0.1)
  * cidr (eg. 192.168.10.0/24)
  * network (eg. 192.168.10.0)
  * netmask (eg. 255.255.255.0)

Example of how to inject routes in interfaces.template:

    {% for ifc in interfaces %}
    auto {{ ifc.name }}
    iface {{ ifc.name }} inet static
    # [...]
    {% for route in ifc.routes %}
        post-up ip route add {{ route.cidr }} via {{ route.gateway }} dev {{ ifc.name }}
        pre-down ip route del {{ route.cidr }} via {{ route.gateway }} dev {{ ifc.name }}
    {% endfor %}
    {% endfor %}

Note: Routes are not injected by default in interfaces.template.
This is to avoid any potential conflicts with existing network agents
such as dhcp, l3, etc.

See nova/tests/network/interfaces-override.template for a complete
example on how to interact with the new interface 'routes' attribute.

Blueprint: network-template-routes-injection
DocImpact: Routes are now available in the network template engine.
Change-Id: Ic22c8af721d539c8180b5caa7ba348e834ddbb56