Updated openstack/openstack Project: openstack-infra/shade  446d4fc5a1c833b709f44238bf3dc56290064322
Add thread sync points to Task

Although the default behavior of Task is to return immediately in shade,
the reason it's abstracted at all is to facilitate multi-threaded
rate-limited TaskManagers such as the one found in nodepool. In order
for those to work, we kind of need the wait conditions to actual do
something.

Change-Id: Icce9729aa2f8f31931dd9d3723d9d5d758436fa9