Updated openstack/openstack Project: openstack-infra/shade  57a5f77de3466c170165c9ff540dcc2834554298
Raise a shade exception on broken volumes

Some clouds don't have volumes and bomb out here, which is a bad place
to bomb out. A later patch will add an active check for service
existence, but we still should catch and throw here so that we're not
leaking python-*client exceptions.

Change-Id: I1f768dcd6ad6f54fb1b5638e9486653ee07965ef