Updated openstack/openstack Project: openstack-infra/project-config  a9159ada1196ee01675f0d6012b594ec64a4310c
Add tempest-dsvm-sahara

Because sahara builds on the compute base layer, it does not need to be
tested in all devstack-gate jobs. Instead create a new job that tests
sahara in a devstack environment as a first step in removing it from the
standard configuration.

Specify that SAHARA should be enabled in the devstack updown job
so that the parts that enable sahara are exercised in devstack.
This is already currently the case, but this updates the job config
so that it will continue to do so later when we remove sahara from
the default config.

Change-Id: Iae4e2439bfdfe7f3fc5f0b75b38583a70d7d78f4
Co-Authored-By: James E. Blair <jeblair@hp.com>