Updated openstack/openstack Project: openstack-infra/project-config  c183c3371090bb21118cf0b059331b8165a8d2a3
Don't install redhat-lsb-core in prepare_node

We don't need it for anything other than detecting whether or not we
should update the sources.list file. Except that's obviously not
something we should do on systems with yum

Remove the call and defer it to install_puppet where it's also
installed.

Change-Id: Ie22aeac1e4c731e7ab61514cb4982c8c35c482e6