Updated openstack/openstack Project: openstack/cinder  92fce883bf12ea1760b80221a0888a35393805f6
Add hacking check for str and unicode in exceptions

One of the comments we are frequently having to make
on reviews is with regards to the use of str() or
unicode() on exceptions.  This hacking check pulled
from Nova will catch this problem earlier and avoid
conflicting comments being made in reviews.

Change-Id: I09eaf7b066f3e29733e6cbe16b2997edb6bc5e23