Updated openstack/openstack Project: openstack/cinder  51ae2ebccd369dbad765178eb058aa42e359f676
Fake out sleeps in unit tests

Just a minor cleanup, had a few tests that were doing sleeps
(or looping calls for those that prefer it) that were being
executed to test timeouts in the unit tests.  No value in
actually *waiting* for these, so this patch just mocks them
out so we still test the procedure but don't actually have
to wait.

This shaves about 30 seconds or so off of the unit test run,
not a huge deal, but every little bit helps IMO.

Also noticed a LOG.debug message in the test_extensions file
that wasn't picked up in other cleanup work so I removed that
debug log call as well.

Change-Id: I415ac75816e8a0f4c48c61b23391df1a9a016d62