Updated openstack/openstack Project: openstack/nova  117351d6c46bd6d00cc92fcd234d67a913833cbc
VMWare: Isolate unit tests from requests

The test_ipv6_host_read actually tested all the way down to the
requests url call, which means it's testing a lot of function
unrelated to the nova code. The interesting part of this test seems to
be that it builds the ipv6 url correctly.

Change this into an assert based on that url being correct. We still
need to mock out the http connection because that's done magically in
__init__.

Unblocks nova patches, which are failing on unit test since the latest
requests release and one of the function signatures changing.

Change-Id: I8ae4cb00e63cbb0304210dba2cd4993016d852ba
Closes-bug: #1451389