Updated openstack/openstack Project: openstack-infra/puppet-log_processor  e28365463580183ca1df30bc2206908a4e59d0c0
Setting dependencies in metadata.json

The log_processor module depends on pip, logstash, and logrotate.
This has implications for the work fungi is doing to delete the pip
module entirely. I think a 'manage_pip' flag to the log_proccessor
module is a good way to conditionally manage pip. OpenstacCI would
leave this off and it would be on by default for users pulling down
this module.

Change-Id: I2038c5436c427243021ed6f578d90d689347120d