Updated openstack/openstack Project: openstack/glance_store  57d9c23677409b1734698720c1e287d9928f44c5
Fix RBD delete image on creation failure

When an exception rises on RBD store while adding/creating an image, and
the image has already been created, this new image is not properly
deleted as it should be.

Fault lies in incorrect call to Store._delete_image method, as it is
called with the wrong arguments.

Change-Id: Ib74f7eafbb04ab893039b480a7d3eaa15e7d59d1
Closes-Bug: #1449639