Updated openstack/openstack Project: openstack-dev/devstack  e210d26a1944ce5e2e8e41201496d93a3d351ddc
clean up logging around run_process

We do a bunch of exec magic unwind in run_process that leads to a lot
of confusing lines in the logs under xtrace. Instead, disable xtrace
through these parts to ensure that the flow at the end of the day
makes more sense.

Change-Id: I91e02465240e704a1a0c0036f5073c0295be018e