Updated openstack/openstack Project: openstack-infra/shade  f525c032a6cb7d2a8451ca6884458ebdafd71571
Switch tasks vs put on a boolean config flag

Rackspace's requirement for task-create is a choice that is orthogonal
to the v2 choice. Make that decision point clear. Also, rename the
hidden methods just so that we're clear on what's happening.

Update to a more recent version of os-client-config, since that also has
support for the task-create flag.

Depends-On: I2cdcfe302a73ebfa7f739399c1eeb3bc9f96ab3c
Change-Id: I72bb5992eeb116f1338a090d997893b9ae93050c