Updated openstack/openstack Project: openstack/openstack-manuals  9a932e0a5c49e6dd804053adfbcdb8b244c9a202
Amendments to the introduction of policy.json

This is an amendment of https://review.openstack.org/#/c/166853/,
after additional information became available.
Specifically, information about the is_admin flag and a note on
tweaking policy.json are added, and some points reworded for
clarity.

While there is still room for improvement, the bug, previously
declared as partially fixed, can be closed now.

Change-Id: Ifd01bc4e367f65d6ce45c54a005bafb660535260
Closes-Bug: #1311067