Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  5be7b768dea17476993916f2aaf8eb2d4323f8e6
Revert "Added parallelization options"

This reverts commit 7100a7f225e60936f23d24c757c1b14c9b136ec4.

The referenced commit, in addition to what it advertises, also
filters the result of update_jobs to include only changed jobs.
If there are no change jobs, then, if the user specifies
--delete-old, there are no jobs in the list of jobs to keep
which comes from the result of update_jobs.  In that case, all
jobs are deleted.  Or in other words, the logic in this change
is that all jobs not updated are deleted.

Proposing as a revert because this is currently causing serious
problems for continuously deployed JJB instances, and the very
common --delete-old code path should be properly tested with
this change before it lands again.

Change-Id: I98443f0c085e27ed8dfece6409434015ac24b306