Updated openstack/openstack Project: openstack/nova  ca658ef29d5b4b77a728c70f42776f9e95f6fc79
Convert NetworkDuplicated to HTTPBadRequest for v2.1 API

If allow_duplicate_networks=False (it is by default) in neutron section
of Nova config, Neutron plugin raises NetworkDuplicated exception for
create instance operation when the same network is specified several
times. This exception is handled by v2 API, but not by v2.1. It leads to
500 internal server error as a result of the call.

This patch adds handling of the exception to return regular
HTTPBadRequest to a user.

Change-Id: Id7d3a3f9725252d999c96ac1ee8e0ff213848c6c