Updated openstack/openstack Project: openstack/cinder  9d1650d5fd175c0d538bf3effe75addfa01d2b33
Logging not using oslo.i18n guidelines (scheduler)

Multi-patch set for easier chunks. This one addresses
the scheduler cinder directory.

There have been quite a few instances found where the
i18n guidelines are not being followed. I believe this
has helped lead to some of the confusion around how to
correctly do this. Other developers see this code and
assume it is an example of the correct usage.

This patch attempts to clean up most of those violations
in the existing codebase to hopefully help avoid some of
that confusion in reviews.

Some issues address:
* Correct log translation markers for different log levels
* Passing format values as arguments to call, not preformatting
* Not forcing translation via six.text_type and others

Guidelines can be found here:
http://docs.openstack.org/developer/oslo.i18n/guidelines.html

Hacking checks will not be able to identify all violations of
the guidelines, but it could be useful for catching obvious
one such as LOG.info("No markers!").

Change-Id: I0473ac402bc7b4a644c9b700df0fa2b88d82d705
Partial-bug: 1433216